Added leaderboard fix
